test refactoring and fixed travis errors
